added custom handshake route for postman testing
